GUI/requirements.txt
cycler==0.10.0
future==0.18.2
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.0
pandas==1.3.0
Pillow==8.3.1
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.1
pytz==2021.1
scipy==1.7.0
six==1.16.0
